Storyboards refactor
- Rename Storyboard to StateManager
- Rename Sheet to ViewState
- Make states stateless, and receive a reference
to their managing state manager in each function
invocation.
