+++ packages/sproutcore/package.json
@@ -10,6 +10,7 @@
     "spade":                 "~> 1.0.0",
     "sproutcore-runtime": "2.0.beta.3",
     "sproutcore-views": "2.0.beta.3",
+    "sproutcore-states": "2.0.beta.3",
     "sproutcore-handlebars": "2.0.beta.3",
     "sproutcore-handlebars-format": "2.0.beta.3"
   },
